fastapi_mongo_auth/requirements.txt
fastapi
uvicorn[standard]
motor
pydantic[email]
passlib[bcrypt]
python-jose[cryptography]
python-multipart
python-dotenv
